requirements.txt
matplotlib
binance-futures-connector
pybit
dataset